backend/requirements.txt
flask
flask-cors
pdfminer.six
requests
python-dotenv
pillow
pytesseract